Chore: merge recent fixes and improvements from dev
- Add reader name for better logging and cache handling
- Add retention API support
- Fix read cache issue
- Fix disk fragment reader load status bug
- Fix fragment manager and log/segment data race
- Improve append latency metrics
- Improve local FS sync handling
- Add log output when closing log file
- Enhance write log details for debugging

+++ common/config/configuration.go
@@ -15,8 +15,8 @@ type MetaConfig struct {
 
 // SegmentRollingPolicyConfig stores the segment rolling policy configuration.
 type SegmentRollingPolicyConfig struct {
-	MaxSize     int `yaml:"maxSize"`
-	MaxInterval int `yaml:"maxInterval"`
+	MaxSize     int64 `yaml:"maxSize"`
+	MaxInterval int   `yaml:"maxInterval"`
 }
 
 type SegmentAppendConfig struct {
@@ -37,13 +37,13 @@ type AuditorConfig struct {
 
 // LogFileSyncPolicyConfig stores the log file sync policy configuration.
 type LogFileSyncPolicyConfig struct {
-	MaxInterval     int `yaml:"maxInterval"`
-	MaxEntries      int `yaml:"maxEntries"`
-	MaxBytes        int `yaml:"maxBytes"`
-	MaxFlushRetries int `yaml:"maxFlushRetries"`
-	RetryInterval   int `yaml:"retryInterval"`
-	MaxFlushSize    int `yaml:"maxFlushSize"`
-	MaxFlushThreads int `yaml:"maxFlushThreads"`
+	MaxInterval     int   `yaml:"maxInterval"`
+	MaxEntries      int   `yaml:"maxEntries"`
+	MaxBytes        int64 `yaml:"maxBytes"`
+	MaxFlushRetries int   `yaml:"maxFlushRetries"`
+	RetryInterval   int   `yaml:"retryInterval"`
+	MaxFlushSize    int64 `yaml:"maxFlushSize"`
+	MaxFlushThreads int   `yaml:"maxFlushThreads"`
 }
 
 // FragmentManagerConfig stores the fragment manager configuration.
